Fix for the dot problem in the class name
If you have a resx file named for example Translations.nl.resx, the module will generate a ts file with a class named Translations.nl which is not a correct class name. With this change the class name will be Translations_nl which is a correct class name and you can continue using the generated ts files as you want.